attestation: move azure generation into attester/, verification into verify.rs
Code motion only - no behavior change, no feature change. Review with
`git diff --color-moved=dimmed-zebra`.

The azure module previously gated generation item by item, scattering
~25 cfg attributes across functions, imports, constants and tests. The
module tree now mirrors the feature lattice instead, so the gate appears
once:

- azure/mod.rs: the shared evidence types (AttestationDocument,
  TpmAttest), payload limits, and MaaError, with the generation-only
  error variants grouped behind one labeled block.
- azure/verify.rs: the verification flow and its tests, compiled
  whenever the module is (azure-verifier is the base feature).
- azure/attester/: everything gated on azure-attester - evidence
  generation, IMDS + vTPM access, AIA certificate fetching, and the
  tss-esapi differential test - behind a single cfg on the mod
  declaration.

The attester module nests inside azure/ rather than sitting beside a
verifier/ peer because the relationship is layered, not symmetric:
generation calls into verification (AIA chain building stops when
verify_ak_cert_with_azure_roots accepts the chain), and there is no
configuration with generation but not verification.

Deltas beyond pure motion, all mechanical: include_bytes paths adjusted
for the new file locations, az_tdx_vtpm::vtpm::Quote abbreviated via the
module's vtpm import, and the differential test reads the quote message
from the fixture JSON instead of TpmQuote's private field.
