+++ crates/attestation/src/azure/mod.rs
@@ -1,44 +1,21 @@
 //! Microsoft Azure vTPM attestation evidence generation and verification
 mod ak_certificate;
 #[cfg(feature = "azure-attester")]
-mod nv_index;
+mod attester;
 mod tpm_quote;
 mod tpms_attest;
+mod verify;
 
 #[cfg(feature = "azure-attester")]
-use std::time::Duration;
-
-use ak_certificate::verify_ak_cert_with_azure_roots;
-#[cfg(feature = "azure-attester")]
-use ak_certificate::{fetch_ak_intermediates_from_aia, read_ak_certificate_from_tpm};
-use az_cvm_vtpm::{hcl, tdx};
-#[cfg(feature = "azure-attester")]
-use az_tdx_vtpm::{imds, vtpm};
-use base64::{Engine as _, engine::general_purpose::URL_SAFE as BASE64_URL_SAFE};
-use dcap_qvl::QuoteCollateralV3;
-use num_bigint::BigUint;
-use openssl::{error::ErrorStack, pkey::PKey};
-use pccs::Pccs;
-#[cfg(feature = "azure-attester")]
-use reqwest::header::CONTENT_TYPE;
+pub use attester::{create_azure_attestation, detect_azure_cvm};
+use az_cvm_vtpm::hcl;
+use openssl::error::ErrorStack;
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use tpm_quote::TpmQuote;
 pub use tpm_quote::TpmQuoteError;
 pub use tpms_attest::AttestError;
-use x509_parser::prelude::*;
-
-use crate::{
-    dcap::{
-        verify_dcap_attestation_with_given_timestamp,
-        verify_dcap_attestation_with_timestamp_sync,
-    },
-    measurements::MultiMeasurements,
-};
-
-/// Used in attestation type detection to check if we are on Azure
-#[cfg(feature = "azure-attester")]
-const AZURE_METADATA_API: &str = "http://169.254.169.254/metadata/instance";
+pub use verify::{get_measurements, verify_azure_attestation, verify_azure_attestation_sync};
 
 /// The attestation evidence payload that gets sent over the channel
 #[derive(Debug, Serialize, Deserialize)]
@@ -128,423 +105,6 @@ where
 
 fn unix_time_now_secs() -> Result<u64, MaaError> {
     Ok(std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH)?.as_secs())
-}
-
-/// Generate a TDX attestation on Azure.
-///
-/// This may perform network calls. Azure's IMDS is queried for the TDX
-/// quote, and the vTPM AK certificate's Authority Information Access (AIA)
-/// CA Issuers URLs are followed to include the observed issuer
-/// intermediates in the evidence.
-///
-/// The intermediates are included as untrusted evidence so verifiers do not
-/// need network access or AIA-fetching logic. This keeps verification
-/// deterministic and easier to reuse in constrained verifier environments
-/// such as TEEs, onchain verification, or zero-knowledge proof generation.
-#[cfg(feature = "azure-attester")]
-pub fn create_azure_attestation(input_data: [u8; 64]) -> Result<Vec<u8>, MaaError> {
-    let hcl_report_bytes = vtpm::get_report_with_report_data(&input_data)?;
-
-    let hcl = hcl::HclReport::new(hcl_report_bytes.clone())?;
-
-    let td_report_from_hcl = hcl.try_into()?;
-
-    // This makes a request to Azure Instance metadata service and gives us a
-    // binary response
-    let td_quote_bytes = imds::get_td_quote(&td_report_from_hcl)?;
-
-    let ak_certificate_der = read_ak_certificate_from_tpm()?;
-    let (remaining_bytes, ak_leaf_certificate) = X509Certificate::from_der(&ak_certificate_der)?;
-    let leaf_len = ak_certificate_der.len() - remaining_bytes.len();
-    let ak_leaf_certificate_der = &ak_certificate_der[..leaf_len];
-    let now_secs = unix_time_now_secs()?;
-    let ak_intermediate_certificates_der =
-        fetch_ak_intermediates_from_aia(ak_leaf_certificate_der, &ak_leaf_certificate, now_secs)?;
-
-    let tpm_attestation = TpmAttest {
-        ak_certificate_pem: pem_rfc7468::encode_string(
-            "CERTIFICATE",
-            pem_rfc7468::LineEnding::default(),
-            ak_leaf_certificate_der,
-        )?,
-        ak_intermediate_certificates_pem: ak_intermediate_certificates_der
-            .iter()
-            .map(|der| {
-                pem_rfc7468::encode_string("CERTIFICATE", pem_rfc7468::LineEnding::default(), der)
-            })
-            .collect::<Result<Vec<_>, _>>()?,
-        quote: (&vtpm::get_quote(&input_data[..32])?).try_into()?,
-        event_log: Vec::new(),
-        instance_info: None,
-    };
-
-    let attestation_document = AttestationDocument {
-        tdx_quote_base64: BASE64_URL_SAFE.encode(&td_quote_bytes),
-        hcl_report_base64: BASE64_URL_SAFE.encode(&hcl_report_bytes),
-        tpm_attestation,
-    };
-
-    tracing::info!("Successfully generated azure attestation: {attestation_document:?}");
-    let attestation_json = serde_json::to_vec(&attestation_document)?;
-    // If this ever fails, then we have a problem and probably just need to
-    // increase MAX_AZURE_ATTESTATION_PAYLOAD_SIZE
-    ensure_azure_attestation_payload_size(&attestation_json)?;
-    Ok(attestation_json)
-}
-
-/// Used during verification to support both sync and async verification
-/// paths without duplicating code
-struct PreparedAzureAttestation {
-    tdx_quote_bytes: Vec<u8>,
-    hcl_report: hcl::HclReport,
-    var_data_hash: [u8; 32],
-    expected_tdx_input_data: [u8; 64],
-    tpm_attestation: TpmAttest,
-}
-
-/// Verify a TDX attestation from Azure
-pub async fn verify_azure_attestation(
-    input: Vec<u8>,
-    expected_input_data: [u8; 64],
-    pccs: Option<Pccs>,
-    override_azure_outdated_tcb: bool,
-) -> Result<MultiMeasurements, MaaError> {
-    let now = unix_time_now_secs()?;
-
-    verify_azure_attestation_with_given_timestamp(
-        input,
-        expected_input_data,
-        pccs,
-        None,
-        now,
-        override_azure_outdated_tcb,
-    )
-    .await
-}
-
-/// Verify a TDX attestation from Azure - synchronous version
-///
-/// This relies on having DCAP collateral already present in the cache
-///
-/// If possible, prefer the async version
-pub fn verify_azure_attestation_sync(
-    input: Vec<u8>,
-    expected_input_data: [u8; 64],
-    pccs: Pccs,
-    override_azure_outdated_tcb: bool,
-) -> Result<MultiMeasurements, MaaError> {
-    let now = unix_time_now_secs()?;
-
-    verify_azure_attestation_with_given_timestamp_sync(
-        input,
-        expected_input_data,
-        pccs,
-        None,
-        now,
-        override_azure_outdated_tcb,
-    )
-}
-
-/// Do the verification, passing in the current time
-/// This allows us to test this function without time checks going out of
-/// date
-async fn verify_azure_attestation_with_given_timestamp(
-    input: Vec<u8>,
-    expected_input_data: [u8; 64],
-    pccs: Option<Pccs>,
-    collateral: Option<QuoteCollateralV3>,
-    now: u64,
-    override_azure_outdated_tcb: bool,
-) -> Result<MultiMeasurements, MaaError> {
-    let PreparedAzureAttestation {
-        tdx_quote_bytes,
-        hcl_report,
-        var_data_hash,
-        expected_tdx_input_data,
-        tpm_attestation,
-    } = prepare_azure_attestation(input)?;
-
-    let _dcap_measurements = verify_dcap_attestation_with_given_timestamp(
-        tdx_quote_bytes,
-        expected_tdx_input_data,
-        pccs,
-        collateral,
-        now,
-        override_azure_outdated_tcb,
-    )
-    .await?;
-
-    finish_azure_attestation_verification(
-        hcl_report,
-        var_data_hash,
-        tpm_attestation,
-        expected_input_data,
-        now,
-    )
-}
-
-/// Synchronous version of the verifier
-fn verify_azure_attestation_with_given_timestamp_sync(
-    input: Vec<u8>,
-    expected_input_data: [u8; 64],
-    pccs: Pccs,
-    collateral: Option<QuoteCollateralV3>,
-    now: u64,
-    override_azure_outdated_tcb: bool,
-) -> Result<MultiMeasurements, MaaError> {
-    let PreparedAzureAttestation {
-        tdx_quote_bytes,
-        hcl_report,
-        var_data_hash,
-        expected_tdx_input_data,
-        tpm_attestation,
-    } = prepare_azure_attestation(input)?;
-
-    let _dcap_measurements = verify_dcap_attestation_with_timestamp_sync(
-        tdx_quote_bytes,
-        expected_tdx_input_data,
-        pccs,
-        collateral,
-        now,
-        override_azure_outdated_tcb,
-    )?;
-
-    finish_azure_attestation_verification(
-        hcl_report,
-        var_data_hash,
-        tpm_attestation,
-        expected_input_data,
-        now,
-    )
-}
-
-/// Parses the attestation during verification
-fn prepare_azure_attestation(input: Vec<u8>) -> Result<PreparedAzureAttestation, MaaError> {
-    ensure_azure_attestation_payload_size(&input)?;
-
-    let attestation_document: AttestationDocument = serde_json::from_slice(&input)?;
-    tracing::info!("Attempting to verify azure attestation: {attestation_document:?}");
-
-    let AttestationDocument { tdx_quote_base64, hcl_report_base64, tpm_attestation } =
-        attestation_document;
-
-    let hcl_report_bytes = BASE64_URL_SAFE.decode(hcl_report_base64)?;
-    let hcl_report = hcl::HclReport::new(hcl_report_bytes)?;
-    let var_data_hash = hcl_report.var_data_sha256();
-
-    let mut expected_tdx_input_data = [0u8; 64];
-    expected_tdx_input_data[..32].copy_from_slice(&var_data_hash);
-
-    let tdx_quote_bytes = BASE64_URL_SAFE.decode(tdx_quote_base64)?;
-
-    Ok(PreparedAzureAttestation {
-        tdx_quote_bytes,
-        hcl_report,
-        var_data_hash,
-        expected_tdx_input_data,
-        tpm_attestation,
-    })
-}
-
-/// The final part of vTPM verification, after verifying DCAP
-fn finish_azure_attestation_verification(
-    hcl_report: hcl::HclReport,
-    var_data_hash: [u8; 32],
-    tpm_attestation: TpmAttest,
-    expected_input_data: [u8; 64],
-    now: u64,
-) -> Result<MultiMeasurements, MaaError> {
-    let hcl_ak_pub = hcl_report.ak_pub()?;
-
-    // Get attestation key from runtime claims
-    let (ak_from_claims, user_data_input) = {
-        let runtime_data_raw = hcl_report.var_data();
-        let claims: HclRuntimeClaims = serde_json::from_slice(runtime_data_raw)?;
-
-        let ak_jwk = claims
-            .keys
-            .iter()
-            .find(|k| k.kid == "HCLAkPub")
-            .ok_or(MaaError::ClaimsMissingHCLAkPub)?;
-
-        let user_data = claims.user_data.as_deref().ok_or(MaaError::ClaimsMissingUserData)?;
-        let user_data_bytes = hex::decode(user_data)?;
-        let user_data_input: [u8; 64] =
-            user_data_bytes.try_into().map_err(|_| MaaError::ClaimsUserDataBadLength)?;
-
-        (RsaPubKey::from_jwk(ak_jwk)?, user_data_input)
-    };
-
-    // Check that the TD report input data matches the HCL var data hash
-    let td_report: tdx::TdReport = hcl_report.try_into()?;
-    if var_data_hash != td_report.report_mac.reportdata[..32] {
-        return Err(MaaError::TdReportInputMismatch);
-    }
-    if user_data_input != expected_input_data {
-        return Err(MaaError::ClaimsUserDataInputMismatch);
-    }
-
-    // Verify the vTPM quote
-    let vtpm_quote = tpm_attestation.quote;
-    let hcl_ak_pub_der = hcl_ak_pub.key.try_to_der().map_err(|_| MaaError::JwkConversion)?;
-    let pub_key = PKey::public_key_from_der(&hcl_ak_pub_der)?;
-    vtpm_quote.verify(&pub_key, &expected_input_data[..32])?;
-
-    let pcrs = vtpm_quote.pcrs_sha256();
-
-    // Parse AK certificate
-    let (_type_label, ak_certificate_der) =
-        pem_rfc7468::decode_vec(tpm_attestation.ak_certificate_pem.as_bytes())?;
-    let ak_intermediate_certificate_ders = tpm_attestation
-        .ak_intermediate_certificates_pem
-        .iter()
-        .map(|pem| pem_rfc7468::decode_vec(pem.as_bytes()).map(|(_type_label, der)| der))
-        .collect::<Result<Vec<_>, _>>()?;
-
-    let (remaining_bytes, ak_certificate) = X509Certificate::from_der(&ak_certificate_der)?;
-    let leaf_len = ak_certificate_der.len() - remaining_bytes.len();
-    let ak_leaf_certificate_der = &ak_certificate_der[..leaf_len];
-
-    // Check that AK public key matches that from TPM quote and HCL claims
-    let ak_from_certificate = RsaPubKey::from_certificate(&ak_certificate)?;
-    let ak_from_hcl = RsaPubKey::from_openssl_pubkey(&pub_key)?;
-    if ak_from_claims != ak_from_hcl {
-        return Err(MaaError::AkFromClaimsNotEqualAkFromHcl);
-    }
-    if ak_from_claims != ak_from_certificate {
-        return Err(MaaError::AkFromClaimsNotEqualAkFromCertificate);
-    }
-
-    // Verify the AK certificate against microsoft root cert
-    verify_ak_cert_with_azure_roots(
-        ak_leaf_certificate_der,
-        &ak_intermediate_certificate_ders,
-        now,
-    )?;
-
-    Ok(MultiMeasurements::from_pcrs(pcrs))
-}
-
-/// Extract the measurements from the attestation, but do not verify
-/// anything. input must be < [MAX_AZURE_ATTESTATION_PAYLOAD_SIZE],
-/// otherwise an error is returned.
-pub fn get_measurements(input: &[u8]) -> Result<MultiMeasurements, MaaError> {
-    ensure_azure_attestation_payload_size(input)?;
-
-    let attestation_document: AttestationDocument = serde_json::from_slice(input)?;
-    let vtpm_quote = attestation_document.tpm_attestation.quote;
-    let pcrs = vtpm_quote.pcrs_sha256();
-    Ok(MultiMeasurements::from_pcrs(pcrs))
-}
-
-/// Detect whether we are on Azure and can make an Azure vTPM attestation
-#[cfg(feature = "azure-attester")]
-pub fn detect_azure_cvm() -> Result<bool, MaaError> {
-    let agent = ureq::AgentBuilder::new().timeout(Duration::from_millis(200)).build();
-    let resp = match agent.get(AZURE_METADATA_API).set("Metadata", "true").call() {
-        Ok(resp) => resp,
-        Err(err) => {
-            tracing::debug!("Azure CVM detection failed: Azure metadata API request failed: {err}");
-            return Ok(false);
-        }
-    };
-
-    if resp.status() != 200 {
-        tracing::debug!(
-            "Azure CVM detection failed: metadata API returned non-success status: {}",
-            resp.status()
-        );
-        return Ok(false);
-    }
-
-    // Ensure the response has a JSON content type
-    let content_type = resp
-        .header(CONTENT_TYPE.as_str())
-        .map(|value| value.to_owned())
-        .ok_or_else(|| MaaError::AzureMetadataApiNonJsonResponse { content_type: None })?;
-
-    if !content_type.to_lowercase().starts_with("application/json") {
-        return Err(MaaError::AzureMetadataApiNonJsonResponse { content_type: Some(content_type) });
-    }
-
-    match az_tdx_vtpm::is_tdx_cvm() {
-        Ok(true) => Ok(true),
-        Ok(false) => {
-            tracing::debug!("Azure CVM detection failed: platform is not an Azure TDX CVM");
-            Ok(false)
-        }
-        Err(err) => {
-            tracing::debug!("Azure CVM detection failed: Azure TDX CVM probe failed: {err}");
-            Ok(false)
-        }
-    }
-}
-
-/// JSON Web Key used in [HclRuntimeClaims]
-#[derive(Debug, Deserialize)]
-struct Jwk {
-    #[allow(unused)]
-    pub kty: String,
-    pub kid: String,
-    #[allow(unused)]
-    pub n: Option<String>,
-    #[allow(unused)]
-    pub e: Option<String>,
-    // other fields ignored
-}
-
-/// The internal data structure for HCL runtime claims
-#[derive(Debug, serde::Deserialize)]
-struct HclRuntimeClaims {
-    keys: Vec<Jwk>,
-    #[allow(unused)]
-    #[serde(rename = "vm-configuration")]
-    vm_config: Option<serde_json::Value>,
-    #[serde(rename = "user-data")]
-    user_data: Option<String>,
-}
-
-/// This is only used as a common type to compare public keys with different
-/// formats
-#[derive(Debug, PartialEq)]
-struct RsaPubKey {
-    n: BigUint,
-    e: BigUint,
-}
-
-impl RsaPubKey {
-    fn from_jwk(jwk: &Jwk) -> Result<Self, MaaError> {
-        if jwk.kty != "RSA" {
-            return Err(MaaError::NotRsa);
-        }
-
-        use base64::engine::general_purpose::URL_SAFE_NO_PAD;
-        let n_bytes = URL_SAFE_NO_PAD.decode(jwk.n.clone().ok_or(MaaError::JwkParse)?)?;
-        let e_bytes = URL_SAFE_NO_PAD.decode(jwk.e.clone().ok_or(MaaError::JwkParse)?)?;
-
-        Ok(Self { n: BigUint::from_bytes_be(&n_bytes), e: BigUint::from_bytes_be(&e_bytes) })
-    }
-
-    fn from_certificate(cert: &X509Certificate) -> Result<Self, MaaError> {
-        let spki = cert.public_key();
-        let Ok(x509_parser::public_key::PublicKey::RSA(rsa_from_cert)) = spki.parsed() else {
-            return Err(MaaError::NotRsa);
-        };
-
-        Ok(Self {
-            n: BigUint::from_bytes_be(rsa_from_cert.modulus),
-            e: BigUint::from_bytes_be(rsa_from_cert.exponent),
-        })
-    }
-
-    fn from_openssl_pubkey(key: &PKey<openssl::pkey::Public>) -> Result<Self, MaaError> {
-        let rsa_from_pkey = key.rsa()?;
-
-        Ok(Self {
-            n: BigUint::from_bytes_be(&rsa_from_pkey.n().to_vec()),
-            e: BigUint::from_bytes_be(&rsa_from_pkey.e().to_vec()),
-        })
-    }
 }
 
 /// An error when generating or verifying a Microsoft Azure vTPM attestation
@@ -642,263 +202,4 @@ pub enum MaaError {
         "Azure metadata API returned a successful response with non-JSON content-type: {content_type:?}"
     )]
     AzureMetadataApiNonJsonResponse { content_type: Option<String> },
-}
-
-#[cfg(all(test, feature = "azure-attester"))]
-mod test_utils {
-    use base64::{Engine as _, engine::general_purpose::URL_SAFE as BASE64_URL_SAFE};
-
-    use super::{AttestationDocument, create_azure_attestation};
-    use crate::dcap::PCS_URL;
-
-    /// Capture a complete Azure TDX attestation fixture from inside an
-    /// Azure TDX CVM.
-    ///
-    /// Run with:
-    ///
-    /// ```text
-    /// cargo test -p attestation --features azure-attester capture_azure_fixture -- --ignored --nocapture
-    /// ```
-    ///
-    /// This writes two timestamped YAML files:
-    ///
-    /// - `azure-tdx-with-ak-intermediates-<timestamp>.yaml`
-    /// - `azure-collateral-with-ak-intermediates-<timestamp>.yaml`
-    ///
-    /// The first file is the full Azure attestation payload, including
-    /// observed vTPM AK intermediate certificates. The second file is
-    /// matching Intel DCAP collateral for the TDX quote embedded in
-    /// that payload, so the fixture can be verified offline by tests.
-    #[tokio::test]
-    #[ignore = "requires an Azure TDX CVM with vTPM access"]
-    async fn capture_azure_fixture() {
-        let output_dir = std::path::Path::new(env!("CARGO_MANIFEST_DIR")).join("test-assets");
-        std::fs::create_dir_all(&output_dir).unwrap();
-
-        // Keep this aligned with existing Azure fixture tests, which use zeroed
-        // report input data.
-        let attestation_json = create_azure_attestation([0u8; 64]).unwrap();
-        let attestation_document: AttestationDocument =
-            serde_json::from_slice(&attestation_json).unwrap();
-
-        let intermediate_count =
-            attestation_document.tpm_attestation.ak_intermediate_certificates_pem.len();
-        assert!(intermediate_count > 0, "captured attestation should include AK intermediates");
-
-        let quote_bytes = BASE64_URL_SAFE.decode(&attestation_document.tdx_quote_base64).unwrap();
-        let quote = dcap_qvl::quote::Quote::parse(&quote_bytes).unwrap();
-        let ca = quote.ca().unwrap();
-        let fmspc = hex::encode_upper(quote.fmspc().unwrap());
-        let collateral = dcap_qvl::collateral::get_collateral_for_fmspc(
-            PCS_URL,
-            fmspc.clone(),
-            ca,
-            false, // TDX, not SGX.
-        )
-        .await
-        .unwrap();
-
-        let timestamp =
-            std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH).unwrap().as_secs();
-        let attestation_path =
-            output_dir.join(format!("azure-tdx-with-ak-intermediates-{timestamp}.yaml"));
-        let collateral_path =
-            output_dir.join(format!("azure-collateral-with-ak-intermediates-{timestamp}.yaml"));
-
-        let mut serializer_options = serde_saphyr::SerializerOptions::default();
-        // With compact list indentation enabled, serde_saphyr can emit an
-        // indentless empty sequence after a nested block sequence, e.g.
-        // `event_log:\n[]`, which its parser rejects. Disable compact list
-        // indentation for fixture output so nested/empty sequences are always
-        // indented under their mapping keys.
-        serializer_options.compact_list_indent = false;
-        std::fs::write(
-            &attestation_path,
-            serde_saphyr::to_string_with_options(&attestation_document, serializer_options)
-                .unwrap(),
-        )
-        .unwrap();
-        std::fs::write(&collateral_path, serde_saphyr::to_string(&collateral).unwrap()).unwrap();
-
-        println!("wrote {}", attestation_path.display());
-        println!("wrote {}", collateral_path.display());
-        println!("quote fmspc={fmspc} ca={ca}");
-        println!("ak_intermediate_certificates_pem entries={intermediate_count}");
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn input_data_from_attestation(attestation_bytes: &[u8]) -> [u8; 64] {
-        let attestation_document: AttestationDocument =
-            serde_saphyr::from_slice(attestation_bytes).unwrap();
-        let hcl_report_bytes =
-            BASE64_URL_SAFE.decode(attestation_document.hcl_report_base64).unwrap();
-        let hcl_report = hcl::HclReport::new(hcl_report_bytes).unwrap();
-        let claims: HclRuntimeClaims = serde_json::from_slice(hcl_report.var_data()).unwrap();
-        let user_data_hex = claims.user_data.unwrap();
-        hex::decode(user_data_hex).unwrap().try_into().unwrap()
-    }
-
-    fn assert_payload_too_large(err: MaaError, actual: usize) {
-        match err {
-            MaaError::AzureAttestationPayloadTooLarge { actual: got, max } => {
-                assert_eq!(got, actual);
-                assert_eq!(max, MAX_AZURE_ATTESTATION_PAYLOAD_SIZE);
-            }
-            err => panic!("unexpected error: {err}"),
-        }
-    }
-
-    #[test]
-    fn get_measurements_rejects_oversized_payload_before_deserialize() {
-        let actual = MAX_AZURE_ATTESTATION_PAYLOAD_SIZE + 1;
-        let input = vec![b'{'; actual];
-
-        let err = match get_measurements(&input) {
-            Ok(_) => panic!("oversized payload should fail"),
-            Err(err) => err,
-        };
-
-        assert_payload_too_large(err, actual);
-    }
-
-    /// All verification entry points must reject an oversized payload, and
-    /// must do so before attempting DCAP verification (no collateral or
-    /// usable PCCS is provided here).
-    #[tokio::test]
-    async fn verify_rejects_oversized_payload_before_deserialize() {
-        let actual = MAX_AZURE_ATTESTATION_PAYLOAD_SIZE + 1;
-        let input = vec![b'{'; actual];
-
-        let err = verify_azure_attestation(input.clone(), [0; 64], None, false).await.unwrap_err();
-        assert_payload_too_large(err, actual);
-
-        let err = verify_azure_attestation_sync(
-            input.clone(),
-            [0; 64],
-            Pccs::new_without_prewarm(None),
-            false,
-        )
-        .unwrap_err();
-        assert_payload_too_large(err, actual);
-
-        let err = verify_azure_attestation_with_given_timestamp(
-            input.clone(),
-            [0; 64],
-            None,
-            None,
-            0,
-            false,
-        )
-        .await
-        .unwrap_err();
-        assert_payload_too_large(err, actual);
-
-        let err = verify_azure_attestation_with_given_timestamp_sync(
-            input,
-            [0; 64],
-            Pccs::new_without_prewarm(None),
-            None,
-            0,
-            false,
-        )
-        .unwrap_err();
-        assert_payload_too_large(err, actual);
-    }
-
-    #[tokio::test]
-    async fn test_decode_hcl() {
-        // From cvm-reverse-proxy/internal/attestation/azure/tdx/testdata/hclreport.
-        // bin
-        let hcl_bytes: &'static [u8] = include_bytes!("../../test-assets/hclreport.bin");
-
-        let hcl_report = hcl::HclReport::new(hcl_bytes.to_vec()).unwrap();
-        let hcl_var_data = hcl_report.var_data();
-        let var_data_values: serde_json::Value = serde_json::from_slice(hcl_var_data).unwrap();
-
-        // Check that it contains 64 byte user data
-        assert_eq!(hex::decode(var_data_values["user-data"].as_str().unwrap()).unwrap().len(), 64);
-    }
-
-    /// Verify a complete observed Azure attestation payload that includes
-    /// AK intermediates fetched from the leaf certificate's AIA URLs.
-    #[tokio::test]
-    async fn test_verify() {
-        // generated using the attester module's [capture_azure_fixture].
-        let attestation_bytes: &'static [u8] =
-            include_bytes!("../../test-assets/azure-tdx-with-ak-intermediates-1780922561.yaml");
-        let collateral_bytes: &'static [u8] = include_bytes!(
-            "../../test-assets/azure-collateral-with-ak-intermediates-1780922561.yaml"
-        );
-
-        // Fixed timestamp within the quote collateral and AK certificate
-        // validity windows, so this offline fixture test does not expire when
-        // wall-clock time advances.
-        let now = 1_780_922_561;
-
-        let attestation_document: AttestationDocument =
-            serde_saphyr::from_slice(attestation_bytes).unwrap();
-        assert_eq!(attestation_document.tpm_attestation.ak_intermediate_certificates_pem.len(), 2);
-
-        let attestation_json = serde_json::to_vec(&attestation_document).unwrap();
-        let async_collateral = serde_saphyr::from_slice(collateral_bytes).unwrap();
-        let sync_collateral = serde_saphyr::from_slice(collateral_bytes).unwrap();
-
-        let async_measurements = verify_azure_attestation_with_given_timestamp(
-            attestation_json.clone(),
-            [0; 64],
-            None,
-            Some(async_collateral),
-            now,
-            false,
-        )
-        .await
-        .unwrap();
-
-        let sync_measurements = verify_azure_attestation_with_given_timestamp_sync(
-            attestation_json,
-            [0; 64],
-            Pccs::new_without_prewarm(None),
-            Some(sync_collateral),
-            now,
-            false,
-        )
-        .unwrap();
-
-        assert_eq!(async_measurements, sync_measurements);
-    }
-
-    #[tokio::test]
-    async fn test_verify_fails_on_input_mismatch() {
-        let attestation_bytes: &'static [u8] =
-            include_bytes!("../../test-assets/azure-tdx-1764662251380464271.yaml");
-        let now = 1771423480;
-
-        let mut expected_input_data = input_data_from_attestation(attestation_bytes);
-        expected_input_data[63] ^= 0x01;
-
-        let collateral_bytes: &'static [u8] =
-            include_bytes!("../../test-assets/azure-collateral02.yaml");
-        let collateral = serde_saphyr::from_slice(collateral_bytes).unwrap();
-        let attestation_json = serde_json::to_vec(
-            &serde_saphyr::from_slice::<AttestationDocument>(attestation_bytes).unwrap(),
-        )
-        .unwrap();
-
-        let err = verify_azure_attestation_with_given_timestamp(
-            attestation_json,
-            expected_input_data,
-            None,
-            Some(collateral),
-            now,
-            false,
-        )
-        .await
-        .unwrap_err();
-
-        assert!(matches!(err, MaaError::ClaimsUserDataInputMismatch));
-    }
 }
